+++ Changelog.md
@@ -5,6 +5,7 @@
 
 - Fixed boolean and datetime handling in transpose method
 - Added the Customers endpoint
+- Added the Addresses endpoint
 
 ## 0.7.0
 ##### 2025-02-12